Refactoring creatTaskElem in a module file
